Increment dc burn metrics for sessions properly
When incrementing the metric, we want the DC to burn metric to be over the cumulative rewardable_dc for all sessions for a pubkey. 

They are inserted into a single row in the DB, so when we grab them back out to burn, we're already dealing with a consolidated value.
