+++ mobile_packet_verifier/tests/integrations/accumulate_sessions.rs
@@ -1,6 +1,3 @@
-use std::str::FromStr;
-
-use anyhow::Context;
 use chrono::{Duration, Utc};
 use file_store::{
     file_sink::{FileSinkClient, MessageReceiver},
@@ -15,9 +12,9 @@ use helium_proto::services::poc_mobile::{
     DataTransferRadioAccessTechnology, VerifiedDataTransferIngestReportV1,
 };
 use mobile_packet_verifier::{
-    accumulate::accumulate_sessions, banning, bytes_to_dc, dc_to_bytes, pending_burns,
+    accumulate::accumulate_sessions, banning, bytes_to_dc, pending_burns,
 };
-use sqlx::{types::Uuid, PgPool};
+use sqlx::PgPool;
 
 use crate::common::{TestChannelExt, TestMobileConfig};
 
@@ -431,102 +428,6 @@ async fn allows_expired_ban_type_data_transfer_keys(pool: PgPool) -> anyhow::Res
     Ok(())
 }
 
-#[sqlx::test]
-async fn small_test(pool: PgPool) -> anyhow::Result<()> {
-    let payer_key =
-        PublicKeyBinary::from_str("112c85vbMr7afNc88QhTginpDEVNC5miouLWJstsX6mCaLxf8WRa")?;
-
-    let (valid_sessions_tx, _valid_sessions_rx) = tokio::sync::mpsc::channel(999_999);
-    let valid_sessions = FileSinkClient::new(valid_sessions_tx, "test");
-    let solana_network = solana::burn::TestSolanaClientMap::default();
-    solana_network.insert(&payer_key, 900_000_000).await;
-
-    let mk_dt = |rewardable_bytes: u64| DataTransferSessionIngestReport {
-        received_timestamp: Utc::now(),
-        report: DataTransferSessionReq {
-            rewardable_bytes,
-            pub_key: PublicKeyBinary::from(vec![1]),
-            signature: vec![],
-            data_transfer_usage: DataTransferEvent {
-                pub_key: PublicKeyBinary::from(vec![1]),
-                upload_bytes: 0,
-                download_bytes: 0,
-                radio_access_technology: DataTransferRadioAccessTechnology::Wlan,
-                event_id: Uuid::new_v4().to_string(),
-                payer: payer_key.clone(),
-                timestamp: Utc::now(),
-                signature: vec![],
-            },
-        },
-    };
-
-    // Fill reports with rewardable_byte values that are just over and under the rounding limit.
-    let mut reports = vec![];
-    for _ in 0..1000 {
-        reports.push(mk_dt(dc_to_bytes(100) + 2));
-        reports.push(mk_dt(dc_to_bytes(150) - 2));
-    }
-
-    let metrics = TestMetrics::new();
-
-    // accumulate and burn
-    run_accumulate_sessions(&pool, reports, TestMobileConfig::all_valid())
-        .await
-        .context("acummulating sessions")?;
-    mobile_packet_verifier::burner::Burner::new(
-        valid_sessions,
-        solana_network.clone(),
-        0,
-        std::time::Duration::default(),
-    )
-    .burn(&pool)
-    .await
-    .context("burning")?;
-
-    metrics.assert_pending_dc_burn(&payer_key, 0).await?;
-
-    Ok(())
-}
-
-#[derive(Clone)]
-struct TestMetrics {
-    addr: String,
-}
-
-impl TestMetrics {
-    fn new() -> Self {
-        let addr = {
-            let listener = std::net::TcpListener::bind("127.0.0.1:0").expect("tcp listener");
-            listener.local_addr().expect("local address")
-        };
-
-        poc_metrics::start_metrics(&poc_metrics::Settings { endpoint: addr })
-            .expect("install prometheus");
-        TestMetrics {
-            addr: format!("http://{addr}"),
-        }
-    }
-
-    async fn assert_pending_dc_burn(
-        &self,
-        payer: &PublicKeyBinary,
-        amount: u64,
-    ) -> anyhow::Result<()> {
-        let res = reqwest::get(self.addr.clone()).await?;
-        let body = res.text().await?;
-        if body.is_empty() {
-            anyhow::bail!("metrics body is empty")
-        }
-
-        let expected = format!(r#"pending_dc_burn{{payer="{}"}} {}"#, payer, amount);
-        if !body.contains(&expected) {
-            anyhow::bail!("expected: {expected} in:\n{body}");
-        }
-
-        Ok(())
-    }
-}
-
 async fn run_accumulate_sessions(
     pool: &PgPool,
     reports: Vec<DataTransferSessionIngestReport>,
@@ -534,7 +435,7 @@ async fn run_accumulate_sessions(
 ) -> anyhow::Result<MessageReceiver<VerifiedDataTransferIngestReportV1>> {
     let mut txn = pool.begin().await?;
 
-    let (verified_sessions_tx, verified_sessions_rx) = tokio::sync::mpsc::channel(999_999);
+    let (verified_sessions_tx, verified_sessions_rx) = tokio::sync::mpsc::channel(10);
     let verified_sessions = FileSinkClient::new(verified_sessions_tx, "test");
 
     let banned_radios = banning::get_banned_radios(&mut txn, Utc::now()).await?;
