+++ mobile_packet_verifier/tests/integrations/main.rs
@@ -2,4 +2,5 @@ pub mod common;
 
 pub mod accumulate_sessions;
 pub mod banning;
+pub mod burn_metric;
 pub mod burner;
